+++ src/components/reportFlow/Analyse.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export function Analyse() {
+	return <>Analyse</>
+}
